Gets the sign in and signup functionality going
sign in and sign up are all set as far as functionality and showing and
hiding when they need to. still need to write the error and success
handling for them. also gets the change password and signout functionality
written but they are both hidden all the time right now as i have to write
the logic to show and hide them appropriately.