Gets all the auth on the UI CRUDing
Set up the sign in and sing up forms together and hide the others pre
sign in. also got the change password and the sign-out forms into the
menu dropdown. Still need to show and hide all of the forms at the right
time and get the auth alerts going.