Builds out the relative position of the manage-market page
Once you are signed in you are now presented with the manage market page
which will allow the bar owner to add beers to their list and start and
stop the market. The page is built out none of the functionality is there.
next task I think is to build the technology or build the market page.