Adds the get ajax call on signin to get beers
created the beers folder in the scripts to handle all the beers actions
this builds out the api call to get all the beers for the current user
that is logged in. next i have to populate the beer list container with
the response that comes back with handle bars.