This adds the save the button to the beer list and hides it
Then when a user clicks on edit the edit and delete buttons hide and the
save button appears. When the user fills out the info and then clicks
save i am grabbing that information and storing it in a beer object. This
is so I can eventually make the ajax request with this data