Builds out the rest of the edit functionality
The edit button now clears the beer name and highlights it and sets the
price to 0.00 on the ui. then the user can update that beer and hit save
which sends a ajax call to patch and then when complete switches the
buttons back to edit and delete and makes the fields un-editable. still
have to build the cancel functionality.