Adds the Druaght Market page
This adds the draught market page and shows it when the user clicks on
the open market button. Added a new template that is almost identical to
the beer list template on the manage market page except for the purchase
button. also hide content on signout and when you click on change password