Builds out the market object with start and stop market methods
Built the market object in the price-logic file to basically take the list
of beers that come back from the get beers server response and put them
into an array so that i can now spit them out into the console over 10 seconds
with the name of the beer and the price. Now i need to get the purchase
button up and running to perform the game-logic.