Builds the purchase beer functionality
built out the purchase beer functionality. I added the logic to do the
calculations on the prices based on what is being bought and what is not.
Also updating the UI with the prices every n about of seconds. I also
changed my logic on edit and on delete beer to re-call get beers after
it successfully edit or deletes respectively. this was just to avoid any
unforeseen bugs that might come from updating the UI and not calling the
db.