Updating a bunch of issues i wanted to clear up
A lot of showing and hiding at the right times. I took the open and close
market buttons off of the fixed position and put them back in the flow of
the website. Fixes game logic defects where it was still running on signout
took away change password on the market page.